Refactor files to classes
