+++ src/EVE/mainThread.py
@@ -1,7 +1,8 @@
 from threading import Thread
+import sys
+sys.path.append("src")
 from windows import *
-from Task import Task
-import time
+from EVETask import EVETask
 
 # todo 
 # restart app when crash?
@@ -9,7 +10,7 @@ import time
 # Keep time difference by 60s all the time. Done partial
 
 def runTask(hwnd, index):
-    task = Task(hwnd, index)
+    task = EVETask(hwnd, index)
     while(True):
         try:
             task.startMiningTask()
@@ -17,8 +18,6 @@ def runTask(hwnd, index):
             task.closeWindow()
             print("thread failed, stop")
             raise(e)
-        
-
 
 def main():
     print("开工前todo list: 打开本地列表,v船到广角,检查快捷列表有无长字符,改变战斗列表为名称排序")
@@ -28,7 +27,6 @@ def main():
         threads.append(Thread(target=runTask, name=str(window["hwnd"]), args=(window["hwnd"],index,)))
     for thread in threads:
         thread.start()
-
 
 
 if __name__ == '__main__':
